requirements.txt
pandas>=2.0.0
neo4j>=5.0.0
python-dotenv>=1.0.0
odfpy>=1.4.1
openpyxl>=3.1.0
groq>=0.4.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
requests>=2.31.0
psycopg2-binary>=2.9.0

